Adds basics card dragging functionality
Adds cards as a scene object so that they can easily be duplicated.
Also adds a Node2D, CardManager, that acts as a basic container for
cards to help maintain scene organization. It also contains the
card_manager script which implements the card mechanics

Adds a Card_Mechanics scene that contains everything so far, I'm
unsure if that will be kept in the future, but for now it helps for
testing card mechanics as they are implemented.