Adds a card Deck that can be clicked to draw cards
This also implements an input manager so that input can control
behavior depending on the collision mask clicked. This means
the input handling from the card_manager script was moved to the
new inputManager script.

This also removes the hand_size variable from player hand as the
Deck script now handles the card instantiation.

The Deck utilizes layer mask 4